Move allocators and handle into common 'mem' namespace and directory
