+++ test/c++/nda_custom_alloc.cpp
@@ -18,7 +18,7 @@
 
 #ifndef _OPENMP
 
-using alloc_t = nda::allocators::segregator<8 * 100, nda::allocators::multiple_bucket<8 * 100>, nda::allocators::mallocator>;
+using alloc_t = nda::mem::segregator<8 * 100, nda::mem::multiple_bucket<8 * 100>, nda::mem::mallocator>;
 
 TEST(CustomAlloc, Create1) { //NOLINT
   nda::basic_array<long, 2, C_layout, 'A', nda::heap_custom_alloc<alloc_t>> A(3, 3);
